# ai-agent-project-frontend/openapi/schema.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8123/api",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/app/update": {
      "post": {
        "tags": [
          "app-controller"
        ],
        "operationId": "updateApp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/app/my/list/page/vo": {
      "post": {
        "tags": [
          "app-controller"
        ],
        "operationId": "listMyAppVOByPage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppQueryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponsePageAppVO"
                }
              }
            }
          }
        }
      }
    },
    "/app/good/list/page/vo": {
      "post": {
        "tags": [
          "app-controller"
        ],
        "operationId": "listGoodAppVOByPage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppQueryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponsePageAppVO"
                }
              }
            }
          }
        }
      }
    },
    "/app/deploy": {
      "post": {
        "tags": [
          "app-controller"
        ],
        "operationId": "deployApp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppDeployDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/app/delete": {
      "post": {
        "tags": [
          "app-controller"
        ],
        "operationId": "deleteApp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/app/admin/update": {
      "post": {
        "tags": [
          "app-controller"
        ],
        "operationId": "updateAppByAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppAdminUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/app/admin/list/page/vo": {
      "post": {
        "tags": [
          "app-controller"
        ],
        "operationId": "listAppVOByPageByAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppQueryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponsePageAppVO"
                }
              }
            }
          }
        }
      }
    },
    "/app/admin/delete": {
      "post": {
        "tags": [
          "app-controller"
        ],
        "operationId": "deleteAppByAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/app/add": {
      "post": {
        "tags": [
          "app-controller"
        ],
        "operationId": "addApp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppAddDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseLong"
                }
              }
            }
          }
        }
      }
    },
...
    "/app/get/vo": {
      "get": {
        "tags": [
          "app-controller"
        ],
        "operationId": "getAppVOById",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseAppVO"
                }
              }
            }
          }
        }
      }
    },
    "/app/download/{appId}": {
      "get": {
        "tags": [
          "app-controller"
        ],
        "operationId": "downloadAppCode",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/app/chat/gen/code": {
      "get": {
        "tags": [
          "app-controller"
        ],
        "operationId": "chatToGenCode",
        "parameters": [
          {
            "name": "appId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "message",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerSentEventString"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/admin/get/vo": {
      "get": {
        "tags": [
          "app-controller"
        ],
        "operationId": "getAppVOByIdByAdmin",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseAppVO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseResponseBoolean": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "BaseResponseLong": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "UserVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userAccount": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "userAvatar": {
            "type": "string"
          },
          "userProfile": {
            "type": "string"
          },
          "userRole": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DeleteRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "AppUpdateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "appName": {
            "type": "string"
          }
        }
      },
      "AppQueryDto": {
        "type": "object",
        "properties": {
          "pageNum": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "sortField": {
            "type": "string"
          },
          "sortOrder": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "appName": {
            "type": "string"
          },
          "cover": {
            "type": "string"
          },
          "initPrompt": {
            "type": "string"
          },
          "codeGenType": {
            "type": "string"
          },
          "deployKey": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AppVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "appName": {
            "type": "string"
          },
          "cover": {
            "type": "string"
          },
          "initPrompt": {
            "type": "string"
          },
          "codeGenType": {
            "type": "string"
          },
          "deployKey": {
            "type": "string"
          },
          "deployedTime": {
            "type": "string",
            "format": "date-time"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/UserVO"
          },
          "editTime": {
            "type": "string",
            "format": "date-time"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BaseResponsePageAppVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/PageAppVO"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "PageAppVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppVO"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int64"
          },
          "pageSize": {
            "type": "integer",
            "format": "int64"
          },
          "totalPage": {
            "type": "integer",
            "format": "int64"
          },
          "totalRow": {
            "type": "integer",
            "format": "int64"
          },
          "optimizeCountQuery": {
            "type": "boolean",
            "writeOnly": true
          }
        }
      },
      "AppDeployDto": {
        "type": "object",
        "properties": {
          "appId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BaseResponseString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "AppAdminUpdateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "appName": {
            "type": "string"
          },
          "cover": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AppAddDto": {
        "type": "object",
        "properties": {
          "initPrompt": {
            "type": "string"
          }
        }
      },
...
      "BaseResponseAppVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/AppVO"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ServerSentEventString": {
        "type": "object"
      }
    }
  },
...
}